refactor(#3264): extract Array.prototype-borrow subsystem from array-methods.ts (#3182)
Verbatim, behaviour-preserving god-file split. Moves the cohesive
`Array.prototype.<method>.call(arrayLike, …)` prototype-borrowing subsystem
(13 symbols: the 4 ARRAY_LIKE_* method sets, standaloneArrayLikeMethodRefused,
compileArrayLikePrototypeCall/Search, compileArrayPrototypeCall and the
specialised borrow impls indexOf/includes/every/some/forEach) out of the
~10.2k-LOC array-methods.ts into a new sibling module
src/codegen/array-prototype-borrow.ts.

- array-methods.ts: 10159 -> 7879 LOC; re-exports the two public entries so
  every existing importer keeps resolving; exports the 6 shared helpers the
  moved code imports back (4 newly exported).
- No logic changes. prove-emit-identity: IDENTICAL (39/39 gc/standalone/wasi
  emits). tsc --noEmit: 0 errors.
- Relocation-shift ratchets: loc-budget / coercion-sites / oracle-ratchet
  granted via change-scoped frontmatter allowances in the issue file (#3131 —
  no whole-tree baseline edits). oracle-ratchet gained the same
  `oracle-ratchet-allow:` frontmatter hatch loc-budget/coercion-sites already
  use, so the allowance lives in this PR's own issue file (no cross-PR churn
  with the sibling god-file-split PRs).
- tests/issue-3264.test.ts smoke-tests the extracted dispatch (7 cases).

✓
